+++ ale-rainbow/ale-rainbow.js
@@ -26,34 +26,39 @@ module.exports = function(RED) {
   }
 
   function releaseSDK(node, instanceId) {
-    node.log ("Removing RainbowSDK instance "+instanceId);
+    node.log ("Removing RainbowSDK (releaseSDK) instance "+instanceId);
     if (rainbowSDK[instanceId] === null) {return};
     delete rainbowSDK[instanceId];
     rainbowSDK[instanceId] = null;
   }
   
   function refreshSDK(node, instanceId) {
-    node.log ("Removing RainbowSDK instance "+instanceId);
+    node.log ("****************************************** IN");
+    node.log ("Removing RainbowSDK (refreshSDK) instance "+instanceId);
     if (rainbowSDK[instanceId] === null) {return};
     stopping = true;
-    // Temporary modification due to issue #4
-    rainbowSDK[instanceId]._core._xmpp.stop(true)
-      .then(() => {
-        return rainbowSDK[instanceId].stop();
-      }).then((res) => {
+    rainbowSDK[instanceId].stop()
+    .catch(function(e) {
+      node.log("catched error during refreshSDK stop : " + JSON.stringify(e)); // "zut !"
+    })
+    .then((res) => 
+      {
+        //removeEventListeners(node, instanceId);
         // SDK has been stopped
         started = false;
         stopping = false;
         stopped = true;
         delete rainbowSDK[instanceId];
         rainbowSDK[instanceId] = null;
+        node.log ("****************************************** OUT");
+/*        
         allocateSDK(node, instanceId);
         // launch it again
         rainbowSDK[instanceId].start().then(() => {
           // SDK is connected to Rainbow
           started = true;
           stopped = false;
-        });
+        });*/
       }
     );
   }
@@ -63,12 +68,14 @@ module.exports = function(RED) {
     if (rainbowSDK[instanceId] === null) {return};
     var handler = rainbowSDKHandlers[instanceId].pop();
      while(handler) {
+       node.log("Remove listenner function :" + handler.fct );
       rainbowSDK[instanceId].events.eee.removeListener(handler.evt, handler.fct);
       handler = rainbowSDKHandlers[instanceId].pop();
      };
   }
 
   function login(config) {
+
     RED.nodes.createNode(this,config);
     var context = this.context();
 
@@ -81,8 +88,9 @@ module.exports = function(RED) {
 
     var successiveCOnnectionFail = 0;
     var node = this;
+    node.log("******************** IN login");
 
-    node.log("Rainbow : login node initialized :"+this)
+    node.log("Rainbow : login node initialized :" + this.id)
 
     SDKoptions[this.id] = {
       rainbow: {
@@ -111,13 +119,14 @@ module.exports = function(RED) {
     };
 
     if (_RainbowSDK === null) {
+      node.log("Rainbow : login node launch require **********");
       _RainbowSDK = require("rainbow-node-sdk");
     }
 
     node.log("config : "+JSON.stringify(config));
     node.log("ENV : http_proxy="+process.env.http_proxy);
     node.log("ENV : https_proxy="+process.env.https_proxy);
-    var ConnectionFail = function () {
+    var ConnectionFail = function ConnectionFail() {
       successiveCOnnectionFail++;
       if (config.proxyHost != "") {
         //We are configure to work behind a proxy.
@@ -140,55 +149,58 @@ module.exports = function(RED) {
     if ((rainbowSDK[node.id] === undefined) || (rainbowSDK[node.id] === null)) {
       allocateSDK (node, node.id);
 
-      var onConnectionReady = function onConnectionReady() {
-        node.log("+++++++++++++++++++");
-        node.log("onConnectionReady()");
-        node.log("+++++++++++++++++++");
+      var onLoginConnectionReady = function onLoginConnectionReady() {
+        node.log("++++++++++++++++++++");
+        node.log("onLoginConnectionReady()");
+        node.log("++++++++++++++++++++");
         successiveCOnnectionFail = 0;
         started = true;
         stopped = stopping = false;
         context.set('RBLoginState',true);
       };
-      rainbowSDK[node.id].events.on('rainbow_onready', onConnectionReady);
-      rainbowSDKHandlers[node.id].push ({evt:'rainbow_onready', fct:onConnectionReady});
+      rainbowSDK[node.id].events.on('rainbow_onready', onLoginConnectionReady);
+      rainbowSDKHandlers[node.id].push ({evt:'rainbow_onready', fct:onLoginConnectionReady});
+      node.log("Rainbow : login register for event 'rainbow_onready'");
 
-      var onConnectionStopped = function onConnectionStopped() {
-        node.log("---------------------");
-        node.log("onConnectionStopped()");
-        node.log("---------------------");
+      var onLoginConnectionStopped = function onLoginConnectionStopped() {
+        node.log("--------------------");
+        node.log("onLoginConnectionStopped()");
+        node.log("--------------------");
         started = false;
         stopped = true;
         stopping = false;
         context.set('RBLoginState',false);
       };
-      rainbowSDK[node.id].events.on('rainbow_onstopped', onConnectionStopped);
-      rainbowSDKHandlers[node.id].push ({evt:'rainbow_onstopped', fct:onConnectionStopped});
+      rainbowSDK[node.id].events.on('rainbow_onstopped', onLoginConnectionStopped);
+      rainbowSDKHandlers[node.id].push ({evt:'rainbow_onstopped', fct:onLoginConnectionStopped});
+      node.log("Rainbow : login register for event 'rainbow_onstopped'");
 
-      var onConnectionError = function onConnectionError( error ) {
+      var onLoginConnectionError = function onLoginConnectionError( error ) {
         var label = ( error ? (error.label ? error.label : error) : "unknown" );
-        node.log("*******************");
-        node.log("onConnectionError() : " + label );
-        node.log("*******************");
+        node.log("********************");
+        node.log("onLoginConnectionError() : " + label );
+        node.log("********************");
         started = false;
         stopped = true;
         stopping = false;
         context.set('RBLoginState',false);
         if (node.autoLogin) {
-          node.log("onConnectionError() - autologin mode");
+          node.log("onLoginConnectionError() - autologin mode");
           ConnectionFail();
           connectTimer = setTimeout(function() {
             refreshSDK(node, node.id);
           },5000);
         }
       }
-      rainbowSDK[node.id].events.on('rainbow_onconnectionerror', onConnectionError);
-      rainbowSDKHandlers[node.id].push ({evt:'rainbow_onconnectionerror', fct:onConnectionError});
+      rainbowSDK[node.id].events.on('rainbow_onconnectionerror', onLoginConnectionError);
+      rainbowSDKHandlers[node.id].push ({evt:'rainbow_onconnectionerror', fct:onLoginConnectionError});
+      node.log("Rainbow : login register for event 'rainbow_onconnectionerror'");
 
-      var onRainbowError = function onRainbowError( error ) {
+      var onLoginRainbowError = function onLoginRainbowError( error ) {
         var label = ( error ? (error.label ? error.label : error) : "unknown" );
-        node.log("////////////////");
-        node.log("onRainbowError() : " + label );
-        node.log("////////////////");
+        node.log("////////////////////");
+        node.log("onLoginRainbowError() : " + label );
+        node.log("////////////////////");
         started = false;
         stopped = true;
         stopping = false;
@@ -200,8 +212,9 @@ module.exports = function(RED) {
           },5000);
         }
       };
-      rainbowSDK[node.id].events.on('rainbow_onerror', onRainbowError);
-      rainbowSDKHandlers[node.id].push ({evt:'rainbow_onerror', fct:onRainbowError});
+      rainbowSDK[node.id].events.on('rainbow_onerror', onLoginRainbowError);
+      rainbowSDKHandlers[node.id].push ({evt:'rainbow_onerror', fct:onLoginRainbowError});
+      node.log("Rainbow : login register for event 'rainbow_onerror'");
 
       if (node.autoLogin) {
         node.log("RainbowSDK instance for autostart "+node.id);
@@ -216,7 +229,10 @@ module.exports = function(RED) {
     }
 
     this.on('close', function(removed, done) {
+      node.log("§§§§§§§§§§§§§§§§§§§§ Node-Red on 'close'- IN");
+
       if (removed) {
+        node.log("Receive 'close' with flag removed set, trying to stop instance " + node.id);
         rainbowSDK[node.id].stop().then(() => {
           // SDK is connected to Rainbow
           started = false;
@@ -224,11 +240,17 @@ module.exports = function(RED) {
           removeEventListeners(node, node.id);
           releaseSDK(node, node.id);
           clearTimeout(connectTimer);
+          node.log("Receive 'close' with flag removed set, instance " + node.id + " stopped and cleared");
+          node.log("§§§§§§§§§§§§§§§§§§§§ Node-Red on 'close'- OUT (removed)");
+          done();
         });        
       } else {
+        node.log("Receive 'close', trying to restart (refresh) instance " + node.id);
+        node.log("§§§§§§§§§§§§§§§§§§§§ Node-Red on 'close'- OUT");
         refreshSDK(node, node.id);
+        done();
       }
-      done();
+      
     });
   }
 
@@ -236,20 +258,20 @@ module.exports = function(RED) {
     RED.nodes.createNode(this,config);
     this.server = RED.nodes.getNode(config.server);
     serverctx = this.server.context();
-    this.filter = config.filter;
+    //this.filter = config.filter;
 
     var cfgTimer = null;
     var node = this;
 
-    node.log("Rainbow : getCnxState node initialized :"+this)
+    node.log("Rainbow : getCnxState node initialized :" + this.id)
 
     node.status({fill:"grey",shape:"dot",text:"off"});
 
     // Update status
-    var getRainbowSDK = function getRainbowSDK() {
+    var getRainbowSDKGetCnxState = function getRainbowSDKGetCnxState() {
       if ((rainbowSDK[config.server] === undefined) || (rainbowSDK[config.server] === null)) {
         node.log("Rainbow SDK not ready ("+config.server+")");
-        cfgTimer = setTimeout (getRainbowSDK, 2000);
+        cfgTimer = setTimeout (getRainbowSDKGetCnxState, 2000);
       } else {
         var onGetRainbowSDKConnectionOk = function onGetRainbowSDKConnectionOk() {
           node.status({fill:"orange",shape:"dot",text:"signin"});
@@ -258,6 +280,7 @@ module.exports = function(RED) {
         };
         rainbowSDK[config.server].events.on('rainbow_onconnected', onGetRainbowSDKConnectionOk);
         rainbowSDKHandlers[config.server].push ({evt:'rainbow_onconnected', fct:onGetRainbowSDKConnectionOk});
+        node.log("Rainbow : getCnxState register for event 'rainbow_onconnected'");
 
         var onGetRainbowSDKConnectionError = function onGetRainbowSDKConnectionError() {
           node.status({fill:"red",shape:"ring",text:"connection error"});
@@ -266,6 +289,7 @@ module.exports = function(RED) {
         };
         rainbowSDK[config.server].events.on('rainbow_onconnectionerror', onGetRainbowSDKConnectionError);
         rainbowSDKHandlers[config.server].push ({evt:'rainbow_onconnectionerror', fct:onGetRainbowSDKConnectionError});
+        node.log("Rainbow : getCnxState register for event 'rainbow_onconnectionerror'");
 
         var onGetRainbowSDKError = function onGetRainbowSDKError() {
           node.status({fill:"red",shape:"ring",text:"config error"});
@@ -274,6 +298,7 @@ module.exports = function(RED) {
         };
         rainbowSDK[config.server].events.on('rainbow_onerror', onGetRainbowSDKError);
         rainbowSDKHandlers[config.server].push ({evt:'rainbow_onerror', fct:onGetRainbowSDKError});
+        node.log("Rainbow : getCnxState register for event 'rainbow_onerror'");
 
         var onGetRainbowSDKReady = function onGetRainbowSDKReady() {
           node.status({fill:"green",shape:"ring",text:"connected"});
@@ -282,13 +307,14 @@ module.exports = function(RED) {
         };
         rainbowSDK[config.server].events.on('rainbow_onready', onGetRainbowSDKReady);
         rainbowSDKHandlers[config.server].push ({evt:'rainbow_onready', fct:onGetRainbowSDKReady});
+        node.log("Rainbow : getCnxState register for event 'rainbow_onready'");
       }
     }
 
     //Eventy from Button to triger Login
     this.on("input",function(msg) {
       node.log ("Got API order "+JSON.stringify(msg));
-      node.log("Rainbow : Login input event: "+JSON.stringify(msg));
+      node.log("Rainbow : GetCnxState input event: "+JSON.stringify(msg));
       try {
         switch (msg.payload) {
           case 'login': {
@@ -309,12 +335,18 @@ module.exports = function(RED) {
           case'logout': {
             node.status({fill:"red",shape:"ring",text:"init"});
             if (started) {
-              rainbowSDK[config.server].stop().then((res) => {
-                // Do something when the SDK has been stopped
-                stopping = false;
-                stopped = true;
-                delete rainbowSDK[instanceId];
-                rainbowSDK[instanceId] = null;
+              // Temporary modification due to issue #4
+              rainbowSDK[config.server]._core._xmpp.stop(true)
+              .then(() => {
+                rainbowSDK[config.server].stop().then((res) => {
+                  // Do something when the SDK has been stopped
+                  stopping = false;
+                  stopped = true;
+                  delete rainbowSDK[config.server];
+                  rainbowSDK[config.server] = null;
+                  var msg = { payload:"stopped by logout" };
+                  node.send(msg);
+                });
               });
             }
             break;
@@ -325,7 +357,7 @@ module.exports = function(RED) {
       }
     });
 
-    getRainbowSDK();
+    getRainbowSDKGetCnxState();
 
     this.on('close', function() {
       // tidy up any state
@@ -333,30 +365,30 @@ module.exports = function(RED) {
     });
   }
 
-    RED.httpAdmin.post(
-      "/Login/:id", 
-      RED.auth.needsPermission("Notified_CnxState.write"), 
-      function(req,res) {
-        var node = RED.nodes.getNode(req.params.id);
-        var state = req.params.state;
-        node.log("Rainbow : Got /Login request: "+JSON.stringify(req.params));
-        if (node != null) {
-          try {
-            var msg = {
-              payload:"login",
-              state:state
-            };
-            node.receive(msg);
-            res.sendStatus(200);
-          } catch(err) {
-            res.sendStatus(500);
-            node.error(RED._("Notified_CnxState.failed",{error:err.toString()}));
-          }
-        } else {
-            res.sendStatus(404);
+  RED.httpAdmin.post(
+    "/Login/:id", 
+    RED.auth.needsPermission("Notified_CnxState.write"), 
+    function(req,res) {
+      var node = RED.nodes.getNode(req.params.id);
+      var state = req.params.state;
+      node.log("Rainbow : Got /Login request: "+JSON.stringify(req.params));
+      if (node != null) {
+        try {
+          var msg = {
+            payload:"login",
+            state:state
+          };
+          node.receive(msg);
+          res.sendStatus(200);
+        } catch(err) {
+          res.sendStatus(500);
+          node.error(RED._("Notified_CnxState.failed",{error:err.toString()}));
         }
+      } else {
+          res.sendStatus(404);
       }
-    );
+    }
+  );
 
   function sendMessage(config) {
     RED.nodes.createNode(this,config);
@@ -366,15 +398,15 @@ module.exports = function(RED) {
     var cfgTimer = null;
     var node = this;
 
-    node.log("Rainbow : sendMessage node initialized :"+this)
+    node.log("Rainbow : sendMessage node initialized :" + this.id)
 
-    var getRainbowSDK = function () {
+    var getRainbowSDKSendMessage = function getRainbowSDKSendMessage() {
       if ((rainbowSDK[config.server] === undefined) || (rainbowSDK[config.server] === null)) {
         node.log("Rainbow SDK not ready ("+config.server+")");
-        cfgTimer = setTimeout (getRainbowSDK, 2000);
+        cfgTimer = setTimeout (getRainbowSDKSendMessage, 2000);
       }
     }
-    getRainbowSDK();
+    getRainbowSDKSendMessage();
 
     this.on('input', function(msg) {
       node.log ("Get msg : "+JSON.stringify(msg));
@@ -407,38 +439,105 @@ module.exports = function(RED) {
   function getMessage(config) {
     RED.nodes.createNode(this,config);
     this.filter = config.filter;
+    this.filterContact = config.filterContact;
+    this.filterCompany = config.filterCompany;
     this.server = RED.nodes.getNode(config.server);
     serverctx = this.server.context();
 
     var cfgTimer = null;
     var node = this;
-    node.log("Rainbow : getMessage node initialized :"+this)
+    node.log("Rainbow : getMessage node initialized :" + this.id)
 
-    var rainbow_onmessagereceived = function(message) {
+    var rainbow_onmessagereceivedGetMessage = function rainbow_onmessagereceivedGetMessage(message) {
+      let filterOK = true;
       node.log("Rainbow : onMessageReceived")
       node.log("Rainbow : Message :"+JSON.stringify(message));
+
+      // RegExp filter ?
       if ((node.filter != '') && (node.filter != undefined)) {
         var regexp = new RegExp(node.filter, 'img');
         node.log("Rainbow : Apply filter :"+node.filter);
-        var res = JSON.stringify(contact).match(regexp);
+        var res = JSON.stringify(message.content).match(regexp);
         node.log("Rainbow : Filter result:"+JSON.stringify(res))
-        if (res == null) return;
+        if (res == null) {
+          filterOK = false;
+          node.log("Rainbow : RegExp filter blocked !");
+          return;
+        }
       }
-      var msg = { payload: message };
-      node.send(msg);
+
+      // Filter with contact/company ?
+      if ( ((node.filterContact != '') && (node.filterContact != undefined)) ||
+        ((node.filterCompany != '') && (node.filterCompany != undefined)) ) {
+
+          node.log("Rainbow : One filter needs to get contact informatopn from jId: " + message.fromJid);
+
+        // Get contact information from server
+        rainbowSDK[config.server]
+//        ._core
+//        ._contacts
+        .contacts
+        .getContactByJid(message.fromJid)
+        .then((contact) => {
+          if (contact) {
+
+            // Contact filter ?
+            if ( (node.filterContact != '') && (node.filterContact != undefined) ) {
+              filterOK = false;
+              if ( (node.filterContact === contact.loginEmail) || (node.filterContact === contact.jid_im) ) {
+                filterOK = true;
+                node.log("Rainbow : Contact filter OK !");
+              } else {
+                node.log("Rainbow : Contact filter blocked !");
+                return;
+              }
+            }
+
+            // Company filter ?
+            if ( (node.filterCompany != '') && (node.filterCompany != undefined) ) {
+              filterOK = false;
+              if (node.filterCompany === contact.companyId) {
+                filterOK = true;
+                node.log("Rainbow : Company id filter OK !");
+              } else {
+                node.log("Rainbow : Company id filter blocked !");
+                return;
+              }
+            }
+
+            node.log("Rainbow : sending message.");
+
+            var msg = { payload: message };
+            node.send(msg);
+
+          } else {
+            node.warn("Rainbow : couldn't find user for jID : " + message.fromJid);
+          }
+
+        });      
+
+      } else {
+        if (!filterOK) {
+          return;
+        }
+        node.log("Rainbow : sending message.");
+        var msg = { payload: message };
+        node.send(msg);
+      }
     };
 
-    var getRainbowSDK = function () {
+    var getRainbowSDKGetMessageGetMessage = function getRainbowSDKGetMessageGetMessage() {
       if ((rainbowSDK[config.server] === undefined) || (rainbowSDK[config.server] === null)) {
         node.log("Rainbow SDK not ready ("+config.server+")");
-        cfgTimer = setTimeout (getRainbowSDK, 2000);
+        cfgTimer = setTimeout (getRainbowSDKGetMessageGetMessage, 2000);
       } else {
         node.log("Rainbow SDK ("+config.server+") Registering rainbow_onmessagereceived");
-        rainbowSDK[config.server].events.on ('rainbow_onmessagereceived',rainbow_onmessagereceived);
-        rainbowSDKHandlers[config.server].push ({evt:'rainbow_onmessagereceived',fct:rainbow_onmessagereceived});
+        rainbowSDK[config.server].events.on ('rainbow_onmessagereceived',rainbow_onmessagereceivedGetMessage);
+        rainbowSDKHandlers[config.server].push ({evt:'rainbow_onmessagereceived',fct:rainbow_onmessagereceivedGetMessage});
+        node.log("Rainbow : getMessage register for event 'rainbow_onmessagereceived'");
       }
     }
-    getRainbowSDK();
+    getRainbowSDKGetMessageGetMessage();
 
     this.on('close', function() {
       // tidy up any state
@@ -457,7 +556,7 @@ module.exports = function(RED) {
 
     node.log("Rainbow : notifyMessageRead node initialized :"+this)
 
-    var rainbow_onmessagereceiptreadreceived = function (message) {
+    var rainbow_onmessagereceiptreadreceived_notifyMessageRead = function (message) {
       node.log("Rainbow : onMessageReceiptReadReceived")
       if ((node.filter != '') && (node.filter != undefined)) {
         var regexp = new RegExp(node.filter, 'img');
@@ -471,17 +570,17 @@ module.exports = function(RED) {
       node.send(msg);
     };
 
-    var getRainbowSDK = function () {
+    var getRainbowSDKnotifyMessageRead = function getRainbowSDKnotifyMessageRead() {
       if ((rainbowSDK[config.server] === undefined) || (rainbowSDK[config.server] === null)) {
         node.log("Rainbow SDK not ready ("+config.server+")");
         cfgTimer = setTimeout (getRainbowSDK, 2000);
       } else {
-        rainbowSDK[config.server].events.on('rainbow_onmessagereceiptreadreceived',rainbow_onmessagereceiptreadreceived);
-        rainbowSDKHandlers[config.server].push ({evt:'rainbow_onmessagereceiptreadreceived',fct:rainbow_onmessagereceiptreadreceived});
+        rainbowSDK[config.server].events.on('rainbow_onmessagereceiptreadreceived',rainbow_onmessagereceiptreadreceived_notifyMessageRead);
+        rainbowSDKHandlers[config.server].push ({evt:'rainbow_onmessagereceiptreadreceived',fct:rainbow_onmessagereceiptreadreceived_notifyMessageRead});
+        node.log("Rainbow : notifyMessageRead register for event 'rainbow_onmessagereceiptreadreceived'");
       }
     }
-    getRainbowSDK();
-
+    getRainbowSDKnotifyMessageRead();
 
     this.on('close', function() {
       // tidy up any state
@@ -501,13 +600,13 @@ module.exports = function(RED) {
 
     node.log("Rainbow : ackMessage node initialized :"+this)
 
-    var getRainbowSDK = function () {
+    var getRainbowSDKackMessage = function () {
       if ((rainbowSDK[config.server] === undefined) || (rainbowSDK[config.server] === null)) {
         node.log("Rainbow SDK not ready ("+config.server+")");
-        cfgTimer = setTimeout (getRainbowSDK, 2000);
+        cfgTimer = setTimeout (getRainbowSDKackMessage, 2000);
       }
     }
-    getRainbowSDK();
+    getRainbowSDKackMessage();
 
     this.on('input', function(msg) {
       node.log ("Ack msg "+JSON.stringify(msg));
@@ -535,7 +634,7 @@ module.exports = function(RED) {
 
     node.log("Rainbow : getContactsPresence node initialized :"+this)
 
-    var rainbow_oncontactpresencechanged = function (contact) {
+    var rainbow_oncontactpresencechanged_getContactsPresence = function rainbow_oncontactpresencechanged_getContactsPresence(contact) {
       node.log("Rainbow : onContactPresenceChanged");
       if ((node.filter != '') && (node.filter != undefined)) {
         var regexp = new RegExp(node.filter, 'img');
@@ -554,16 +653,17 @@ module.exports = function(RED) {
       clearTimeout (cfgTimer);
     });
 
-    var getRainbowSDK = function () {
+    var getRainbowSDKgetContactsPresence = function getRainbowSDKgetContactsPresence() {
       if ((rainbowSDK[config.server] === undefined) || (rainbowSDK[config.server] === null)) {
         node.log("Rainbow SDK not ready ("+config.server+")");
-        cfgTimer = setTimeout (getRainbowSDK, 2000);
+        cfgTimer = setTimeout (getRainbowSDKgetContactsPresence, 2000);
       } else {
-        rainbowSDK[config.server].events.on('rainbow_oncontactpresencechanged',rainbow_oncontactpresencechanged);
-        rainbowSDKHandlers[config.server].push ({evt:'rainbow_oncontactpresencechanged',fct:rainbow_oncontactpresencechanged});
+        rainbowSDK[config.server].events.on('rainbow_oncontactpresencechanged',rainbow_oncontactpresencechanged_getContactsPresence);
+        rainbowSDKHandlers[config.server].push ({evt:'rainbow_oncontactpresencechanged',fct:rainbow_oncontactpresencechanged_getContactsPresence});
+        node.log("Rainbow : getContactsPresence register for event 'rainbow_oncontactpresencechanged'");
       }
     }
-    getRainbowSDK();
+    getRainbowSDKgetContactsPresence();
   }
 
 
@@ -576,16 +676,16 @@ module.exports = function(RED) {
     var cfgTimer = null;
     var node = this;
 
-    node.log("Rainbow : setPresence node initialized :"+this)
+    node.log("Rainbow : setPresence node initialized :" + this.id);
 
-    var getRainbowSDK = function () {
+    var getRainbowSDKsetPresence = function getRainbowSDKsetPresence() {
       if ((rainbowSDK[config.server] === undefined) || (rainbowSDK[config.server] === null)) {
         node.log("Rainbow SDK not ready ("+config.server+")");
-        cfgTimer = setTimeout (getRainbowSDK, 2000);
+        cfgTimer = setTimeout (getRainbowSDKsetPresence, 2000);
       } else {
       }
     }
-    getRainbowSDK();
+    getRainbowSDKsetPresence();
 
     this.on('input', function(msg) {
       node.log ("Set presence "+JSON.stringify(msg));